+++ shared/src/main/java/mobi/nowtechnologies/server/shared/dto/web/PaymentHistoryItemDto.java
@@ -22,7 +22,7 @@ public class PaymentHistoryItemDto {
 
 	private String paymentMethod;
 
-	private long duration;
+	private int duration;
 
 	private DurationUnit durationUnit;
 
@@ -70,7 +70,7 @@ public class PaymentHistoryItemDto {
 		return duration;
 	}
 
-	public void setDuration(long duration) {
+	public void setDuration(int duration) {
 		this.duration = duration;
 	}
